refactor(api): give PATCH /tasks/:taskId a domain module
The handler held one complete domain action inline: ~78 lines of preflight
refusals reading `db` five times, then four `$transaction` calls. All of it
was closure-independent, so it moves whole.

`patchTask(db, taskId, patch)` owns the preflight reads, all four
transactions and their isolation level, and returns either the written task
or a `TaskPatchRefusal`. The 13 inline `context.json(...)` refusals become
returned values the route maps once.

The four transactions stay four: three are guarded by different conditions
and fire on different requests.

`isLiveStatus`, `hasActiveRun` and `reactivationBlocked` move to
`task-write.ts` — they are shared with routes that stay in `app.ts`, and the
lock-order protocol they belong to already lives there. `withoutUndefined`
gets its own file for the same reason. `taskWriteRefusal`, `StatusWritePlan`
and `GateWinner` had no other caller and move into the module.
